backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
yfinance==0.2.37
pandas==2.1.4
ta
numpy==1.26.3
scikit-learn==1.4.0
tensorflow==2.15.0
hmmlearn==0.3.0
transformers==4.37.0
torch==2.1.2
praw==7.7.1
tweepy==4.14.0
newsapi-python==0.2.7
pytrends==4.9.2
requests==2.31.0
beautifulsoup4==4.12.3
aiohttp==3.9.1
python-multipart==0.0.6
httpx==0.26.0
textblob==0.18.0
